Do not manipulate Apache in Nginx module
Previously, we were stopping Apache service and removing
its packages.  This was a very dangerous behavior, as
Apache and Nginx can (and often do) co-exist on the same
system, running on different ports.

We don't do that anymore.  Nginx module is only trying to
install and start Nginx with PHP-FPM.  If Apache is already
running on the same port, Nginx will fail to start and be
done with it.  Which is a way better alternative to breaking
the working installation.