More elegant solution for backup dir (task #355)
Puppet doesn't know how to create directories recursively, so
there is no elegant solution to this problem currently.  However,
since we ship the backup running shell script, we can rely on
the "mkdir -p" to create the folder before we do the backup
there.  Access rights should be sufficient as well.